modified:   build/libcalculix_plugin.so 	new file:   src/Core/CoreResultsFrd copy.cpp 	new file:   src/Core/CoreResultsFrd copy.hpp 	modified:   src/Core/CoreResultsFrd.cpp 	modified:   src/Core/CoreResultsFrd.hpp
